Rename folder index files to route files
This feels a bit more natural and a better naming convention to have
